pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "guided-dc"
version = "0.1.0"
description = "Guided data collection."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    # Maniskill dependencies
    "numpy>=1.22,<2.0.0",
    "scipy",
    "dacite",
    "gymnasium==0.29.1",
    "sapien==3.0.0.b1",
    "h5py",
    "pyyaml",
    "tqdm",
    "GitPython",
    "tabulate",
    "transforms3d",
    "trimesh",
    "imageio",
    "imageio[ffmpeg]",
    "mplib==0.1.1;platform_system=='Linux'",
    "fast_kinematics==0.2.2;platform_system=='Linux'",
    "IPython",
    "pytorch_kinematics_ms==0.7.3",  # pytorch kinematics package for ManiSkill forked from https://github.com/UM-ARM-Lab/pytorch_kinematics
    "tyro>=0.8.5",  # nice, typed, command line arg parser
    "huggingface_hub",  # we use HF to version control some assets/datasets more easily
    "torch",
    "torchvision",
    "torchaudio",
    # Guided-dc dependencies
    "omegaconf==2.3.0",
    "hydra-core==1.3.2",
    "moviepy==1.0.3",
    # DPPO dependencies
    "timm",
]

[tool.setuptools.packages.find]
exclude = []

[tool.wheel]
exclude = []

# Maniskill
[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "serial",
    "gpu_sim: marks tests as using the GPU simulation (deselect with '-m \"not gpu_sim\"')",
    "serial",
]